Use Terraform Registry download URL (#3)
Use Terraform Registry download URL

The version tags used on the Terraform Registry don't necessarily match those in the repo, e.g. `0.7.0` compared to `v0.7.0`. This change uses the Terraform Registry download URL to determine the GitHub download URL, and grabs the repo version tag from that.

Another change is to make local filesystem modules only work if they start with certain characters, and to not check if they exist or not. If they don't exist, it will error. The paths must also be relative to the Terrafile (pterrafile can run from a different directory, so behaviour would have been inconsistent).